+++ src/components/transaction-address.tsx
@@ -107,7 +107,7 @@ export const TransactionAddress = ({
     <KeyboardSafeArea>
       <TextField
         label={getText(I18N.transactionAddressLabel)}
-        style={page.input}
+        style={styles.input}
         value={address}
         onChangeText={setAddress}
         error={error}
@@ -118,7 +118,7 @@ export const TransactionAddress = ({
         placeholder={getText(I18N.transactionAddressPlaceholder)}
         rightAction={
           address === '' ? (
-            <View style={page.inputButtonContainer}>
+            <View style={styles.inputButtonContainer}>
               <IconButton onPress={onPressPaste}>
                 <Icon i24 name="paste" color={Color.graphicGreen1} />
               </IconButton>
@@ -144,13 +144,13 @@ export const TransactionAddress = ({
         variant={ButtonVariant.contained}
         title={getText(I18N.continue)}
         onPress={onDone}
-        style={page.button}
+        style={styles.button}
       />
     </KeyboardSafeArea>
   );
 };
 
-const page = StyleSheet.create({
+const styles = StyleSheet.create({
   input: {
     marginBottom: 12,
     marginHorizontal: 20,
