+++ src/components/ui/swipeable-row.tsx
@@ -1,10 +1,9 @@
 import React, {useEffect, useMemo, useRef} from 'react';
 
-import {Animated, View} from 'react-native';
+import {Animated} from 'react-native';
 import {Swipeable} from 'react-native-gesture-handler';
 
-import {SwipeableButton} from '@app/components/ui/swipeable-button';
-import {createTheme} from '@app/helpers';
+import {SwipeableButtonList} from '@app/components/swipeable-button-list';
 import {SwipeableAction} from '@app/types';
 
 export type SwipeableRowProps<T> = {
@@ -27,19 +26,11 @@ export const SwipeableRow = ({
   const rActions = useMemo(
     () => (progress: Animated.AnimatedInterpolation<number>) =>
       (
-        <View style={[styles.container, {width: rightActions.length * 74}]}>
-          {rightActions.map((action, i) => (
-            <SwipeableButton
-              backgroundColor={action.backgroundColor}
-              progress={progress}
-              x={(rightActions.length - i) * 74}
-              key={action.key}
-              item={item}
-              icon={action.icon}
-              onPress={action.onPress}
-            />
-          ))}
-        </View>
+        <SwipeableButtonList
+          item={item}
+          rightActions={rightActions}
+          progress={progress}
+        />
       ),
     [item, rightActions],
   );
@@ -50,9 +41,3 @@ export const SwipeableRow = ({
     </Swipeable>
   );
 };
-
-const styles = createTheme({
-  container: {
-    flexDirection: 'row',
-  },
-});
